requirements.txt
asgiref==3.11.0
Django==5.2.9
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
rest-framework-simplejwt==0.0.2
sqlparse==0.5.5
typing_extensions==4.15.0
